Update BaseCamp_7K.CR1
Crane,

Here are some edits to the program. Mostly just mods to commenting for clarity, but I corrected an error in the wiring for the wind monitor  and changed the wiring for the CS215 back to 12V, pending the required change in the program that turns on the switched power so we can use the SW12V wiring.